+++ tests/IntegrationTests/DI/InjectionTest.php
@@ -9,11 +9,14 @@
 
 namespace IntegrationTests\DI;
 
+use DI\ContainerBuilder;
 use DI\Definition\FileLoader\ArrayDefinitionFileLoader;
 use DI\Definition\FileLoader\YamlDefinitionFileLoader;
 use DI\Scope;
 use DI\Container;
 use IntegrationTests\DI\Fixtures\Class1;
+use IntegrationTests\DI\Fixtures\Class2;
+use IntegrationTests\DI\Fixtures\Implementation1;
 use IntegrationTests\DI\Fixtures\LazyDependency;
 
 /**
@@ -36,9 +39,10 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
     public static function containerProvider()
     {
         // Test with a container using reflection
-        $containerReflection = new Container();
-        $containerReflection->useReflection(true);
-        $containerReflection->useAnnotations(false);
+        $builder = new ContainerBuilder();
+        $builder->useReflection(true);
+        $builder->useAnnotations(false);
+        $containerReflection = $builder->build();
         $containerReflection->addDefinitions(
             array(
                 'foo'                                     => 'bar',
@@ -52,9 +56,10 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
         );
 
         // Test with a container using annotations and reflection
-        $containerAnnotations = new Container();
-        $containerAnnotations->useReflection(true);
-        $containerAnnotations->useAnnotations(true);
+        $builder = new ContainerBuilder();
+        $builder->useReflection(true);
+        $builder->useAnnotations(true);
+        $containerAnnotations = $builder->build();
         $containerAnnotations->addDefinitions(
             array(
                 'foo'                                     => 'bar',
@@ -68,16 +73,18 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
         );
 
         // Test with a container using array configuration
-        $containerArray = new Container();
-        $containerArray->useReflection(false);
-        $containerArray->useAnnotations(false);
+        $builder = new ContainerBuilder();
+        $builder->useReflection(false);
+        $builder->useAnnotations(false);
+        $containerArray = $builder->build();
         $array = require __DIR__ . '/Fixtures/definitions.php';
         $containerArray->addDefinitions($array);
 
         // Test with a container using PHP configuration
-        $containerPHP = new Container();
-        $containerPHP->useReflection(false);
-        $containerPHP->useAnnotations(false);
+        $builder = new ContainerBuilder();
+        $builder->useReflection(false);
+        $builder->useAnnotations(false);
+        $containerPHP = $builder->build();
         $containerPHP->set('foo', 'bar');
         $containerPHP->set('IntegrationTests\DI\Fixtures\Class1')
             ->withScope(Scope::PROTOTYPE())
@@ -109,16 +116,18 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
         $containerPHP->set('IntegrationTests\DI\Fixtures\LazyDependency');
 
         // Test with a container using array configuration loaded from file
-        $containerArrayFromFile = new Container();
-        $containerArrayFromFile->useReflection(false);
-        $containerArrayFromFile->useAnnotations(false);
-        $containerArrayFromFile->addDefinitionsFromFile(new ArrayDefinitionFileLoader(__DIR__ . '/Fixtures/definitions.php'));
+        $builder = new ContainerBuilder();
+        $builder->useReflection(false);
+        $builder->useAnnotations(false);
+        $builder->addDefinitionsFromFile(new ArrayDefinitionFileLoader(__DIR__ . '/Fixtures/definitions.php'));
+        $containerArrayFromFile = $builder->build();
 
         // Test with a container using array configuration loaded from file
-        $containerYaml = new Container();
-        $containerYaml->useReflection(false);
-        $containerYaml->useAnnotations(false);
-        $containerYaml->addDefinitionsFromFile(new YamlDefinitionFileLoader(__DIR__ . '/Fixtures/definitions.yml'));
+        $builder = new ContainerBuilder();
+        $builder->useReflection(false);
+        $builder->useAnnotations(false);
+        $builder->addDefinitionsFromFile(new YamlDefinitionFileLoader(__DIR__ . '/Fixtures/definitions.yml'));
+        $containerYaml = $builder->build();
 
         return array(
             array(self::DEFINITION_REFLECTION, $containerReflection),
@@ -137,6 +146,7 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
     {
         /** @var $class1 Class1 */
         $class1 = $container->get('IntegrationTests\DI\Fixtures\Class1');
+
         $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Class2', $class1->constructorParam1);
         $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Implementation1', $class1->constructorParam2);
     }
@@ -152,6 +162,35 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
         }
         /** @var $class1 Class1 */
         $class1 = $container->get('IntegrationTests\DI\Fixtures\Class1');
+
+        $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Class2', $class1->property1);
+        $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Implementation1', $class1->property2);
+        $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Class2', $class1->property3);
+        $this->assertEquals('bar', $class1->property4);
+        // Lazy injection
+        /** @var LazyDependency|\ProxyManager\Proxy\LazyLoadingInterface $proxy */
+        $proxy = $class1->property5;
+        $this->assertInstanceOf('IntegrationTests\DI\Fixtures\LazyDependency', $proxy);
+        $this->assertInstanceOf('ProxyManager\Proxy\LazyLoadingInterface', $proxy);
+        $this->assertFalse($proxy->isProxyInitialized());
+        // Correct proxy resolution
+        $this->assertTrue($proxy->getValue());
+        $this->assertTrue($proxy->isProxyInitialized());
+    }
+
+    /**
+     * @dataProvider containerProvider
+     */
+    public function testPropertyInjectionExistingObject($type, Container $container)
+    {
+        // Only constructor injection with reflection
+        if ($type == self::DEFINITION_REFLECTION) {
+            return;
+        }
+        /** @var $class1 Class1 */
+        $class1 = new Class1(new Class2(), new Implementation1());
+        $container->injectOn($class1);
+
         $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Class2', $class1->property1);
         $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Implementation1', $class1->property2);
         $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Class2', $class1->property3);
@@ -178,6 +217,35 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
         }
         /** @var $class1 Class1 */
         $class1 = $container->get('IntegrationTests\DI\Fixtures\Class1');
+
+        $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Class2', $class1->method1Param1);
+        $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Implementation1', $class1->method2Param1);
+        $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Class2', $class1->method3Param1);
+        $this->assertEquals('bar', $class1->method3Param2);
+        $this->assertInstanceOf('IntegrationTests\DI\Fixtures\LazyDependency', $class1->method4Param1);
+        $this->assertInstanceOf('ProxyManager\Proxy\LazyLoadingInterface', $class1->method4Param1);
+        // Lazy injection
+        /** @var LazyDependency|\ProxyManager\Proxy\LazyLoadingInterface $proxy */
+        $proxy = $class1->method4Param1;
+        $this->assertFalse($proxy->isProxyInitialized());
+        // Correct proxy resolution
+        $this->assertTrue($proxy->getValue());
+        $this->assertTrue($proxy->isProxyInitialized());
+    }
+
+    /**
+     * @dataProvider containerProvider
+     */
+    public function testMethodInjectionExistingObject($type, Container $container)
+    {
+        // Only constructor injection with reflection
+        if ($type == self::DEFINITION_REFLECTION) {
+            return;
+        }
+        /** @var $class1 Class1 */
+        $class1 = new Class1(new Class2(), new Implementation1());
+        $container->injectOn($class1);
+
         $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Class2', $class1->method1Param1);
         $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Implementation1', $class1->method2Param1);
         $this->assertInstanceOf('IntegrationTests\DI\Fixtures\Class2', $class1->method3Param1);